+++ examples/crisp.go
@@ -16,11 +16,11 @@ func main() {
 
 	// minor grid
 	for x := Minor; x < W; x += Minor {
-		fx := float64(x) + 0.5
+		fx := float32(x) + 0.5
 		dc.DrawLine(fx, 0, fx, H)
 	}
 	for y := Minor; y < H; y += Minor {
-		fy := float64(y) + 0.5
+		fy := float32(y) + 0.5
 		dc.DrawLine(0, fy, W, fy)
 	}
 	dc.SetLineWidth(1)
@@ -29,11 +29,11 @@ func main() {
 
 	// major grid
 	for x := Major; x < W; x += Major {
-		fx := float64(x) + 0.5
+		fx := float32(x) + 0.5
 		dc.DrawLine(fx, 0, fx, H)
 	}
 	for y := Major; y < H; y += Major {
-		fy := float64(y) + 0.5
+		fy := float32(y) + 0.5
 		dc.DrawLine(0, fy, W, fy)
 	}
 	dc.SetLineWidth(1)
