+++ examples/openfill.go
@@ -11,8 +11,8 @@ func main() {
 	dc := gg.NewContext(1000, 1000)
 	for j := 0; j < 10; j++ {
 		for i := 0; i < 10; i++ {
-			x := float64(i)*100 + 50
-			y := float64(j)*100 + 50
+			x := float32(i)*100 + 50
+			y := float32(j)*100 + 50
 			a1 := rand.Float64() * 2 * math.Pi
 			a2 := a1 + rand.Float64()*math.Pi + math.Pi/2
 			dc.DrawArc(x, y, 40, a1, a2)
